requirement.txt
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.0.1
PyJWT==2.3.0
djangorestframework-simplejwt==5.2.2